Updated the AdminContainer to accept component rather than child and updated all the admin routes to accomodate the change
