+++ src/Components/admin/editAnnouncements/index.jsx
@@ -1,36 +1,19 @@
 import React from 'react';
-import { useQuery } from '@apollo/react-hooks';
 import { useParams } from 'react-router-dom';
-import AnnouncementEditor from './AnnouncementEditor';
-import { GET_CONTEST_ANNOUNCEMENT } from '../../../graphql/queries';
-import Spinner from '../../common/Spinner/index';
-import SomethingWentWrong from '../../common/SomethingWentWrong/index';
 import AdminContainer from '../AdminContainer';
-import useSentry from '../../../customHooks/useSentry';
+import EditAnnoucements from './EditAnnoucements';
 
-const EditAnnouncements = () => {
+const EditAnnoucementsPage = () => {
   const { contestId } = useParams();
-  const { logError } = useSentry();
-  const { loading, error, data } = useQuery(GET_CONTEST_ANNOUNCEMENT, {
-    variables: { code: contestId },
-  });
-  if (loading) return <Spinner />;
-  if (error) {
-    logError('adminDashboad query in EditAnnoucements', { ...data, ...error });
-    return <SomethingWentWrong message="An error has been encountered." />;
-  }
-  if (data.adminDashboard
-    && data.adminDashboard.contest) {
-    return (
-      <AdminContainer contestCode={contestId}>
-        <AnnouncementEditor announcement={data.adminDashboard.contest.announcement ? data.adminDashboard.contest.announcement : ''} />
-      </AdminContainer>
-    );
-  }
-
-  // random errors not handled by graphql
-  logError('adminDashboad query in EditAnnoucements', { ...data, ...error });
-  return <SomethingWentWrong message="An unexpected error has been encountered" />;
+  // AdminContainer will check whether the user is admin or not and
+  // if the user is admin only then the user will be allowed to see the
+  // component
+  return (
+    <AdminContainer
+      component={<EditAnnoucements />}
+      contestCode={contestId}
+    />
+  );
 };
 
-export default EditAnnouncements;
+export default EditAnnoucementsPage;
